second iteration of file validation. Non-happy path cases accounted for, previous image persisting in the event of failed uploads

+++ App/Lib/Helpers/MediaHandler.php
@@ -4,17 +4,17 @@ class MediaHandler
 {
     public static function imageValidator(): string
     {
-        print_r($_FILES);
-        $validMimes=['image/jpeg','image/png'];
         $errorMessage = "";
-        //conditional logic inspection of file.
-        if ($_FILES['perf_img']['size'] >= 100000) {
-            //upload_max_filesize=40M
-            $errorMessage = 'file must be under 1MB';
-        } elseif (strlen($_FILES['perf_img']['name']) >= 20) {
-            $errorMessage = 'file name must be under 20 characters';
-        } elseif (!in_array(mime_content_type($_FILES['perf_img']['tmp_name']), $validMimes)) {
-            $errorMessage = 'file name must be of type image/jpeg, image/png';
+        if (!empty($_FILES['perf_img']['name'])) {
+            $validMimes=['image/jpeg','image/png'];
+            if ($_FILES['perf_img']['size'] >= 100000) {
+                //upload_max_filesize=40M
+                $errorMessage = 'file must be under 1MB';
+            } elseif (strlen($_FILES['perf_img']['name']) >= 20) {
+                $errorMessage = 'file name must be under 20 characters';
+            } elseif (!in_array(mime_content_type($_FILES['perf_img']['tmp_name']), $validMimes)) {
+                $errorMessage = 'file name must be of type image/jpeg, image/png';
+            }
         }
         return $errorMessage;
     }
@@ -30,9 +30,11 @@ class MediaHandler
             };
         }
         $prevImage = scandir($imageUploadDir, 1);
-        var_dump($prevImage);
-        if (count($prevImage)>2) {
+        if (count($prevImage)>2 && !empty($_FILES['perf_img']['name'])) {
             unlink($imageUploadDir.$prevImage[0]);
+        } elseif (count($prevImage)>2 && empty($_FILES['perf_img']['name'])) {
+            $currentImageRelPath = '/' . $id . '/' . $prevImage[0];
+            return $currentImageRelPath;
         }
         if (!move_uploaded_file($_FILES['perf_img']['tmp_name'], $imageUploadFile)) {
             echo 'failed to move file!';
